+++ src/modules/rocket/active_launches_window.cpp
