+++ src/modules/rocket/launch_window.cpp
