+++ src/modules/site/building_window.cpp
@@ -3,9 +3,9 @@
 #include "modules/base/assert.h"
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
-#include "modules/rocket_launch/actions.h"
-#include "modules/rocket_launch/launch_window.h"
-#include "modules/rocket_launch/rocket_launch.h"
+#include "modules/rocket/actions.h"
+#include "modules/rocket/launch_window.h"
+#include "modules/rocket/rocket_launch.h"
 #include "modules/stats/stats.h"
 #include "rocket_prefab_window.h"
 #include "site.h"
