+++ test/active_launches_tests.cpp
@@ -1,5 +1,5 @@
-#include "modules/rocket_launch/active_launches_window.h"
-#include "modules/rocket_launch/rocket_launch.h"
+#include "modules/rocket/active_launches_window.h"
+#include "modules/rocket/rocket_launch.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
 #include <catch2/catch_test_macros.hpp>
@@ -29,9 +29,9 @@ struct ActiveLaunchesFixture {
   flecs::entity plan3; // pad1, no payload
 
   ActiveLaunchesFixture() {
-    world.import <SimulationModule>();
-    world.import <SiteModule>();
-    world.import <RocketLaunchModule>();
+    world.import<SimulationModule>();
+    world.import<SiteModule>();
+    world.import<RocketLaunchModule>();
 
     site1 = world.entity("Site A").add<Site>();
     site2 = world.entity("Site B").add<Site>();
