+++ test/rocket_prefab_window.test.cpp
@@ -1,5 +1,5 @@
 #include "modules/site/rocket_prefab_window.h"
-#include "modules/rocket_launch/rocket_launch.h"
+#include "modules/rocket/rocket_launch.h"
 #include "modules/simulation/simulation.h"
 #include "modules/stats/stats.h"
 #include <catch2/catch_test_macros.hpp>
